ENH: sparse/dsolve: add unmodified SuperLU 4.3 sources
Except:

- Removed *myblas2.c (we USE_VENDOR_BLAS)

- Removed mc64ad.* (from Harwell Subroutine Library, license issues)

+++ scipy/sparse/linalg/dsolve/SuperLU/SRC/memory.c
@@ -133,7 +133,7 @@ user_bcopy(char *src, char *dest, int bytes)
 int *intMalloc(int n)
 {
     int *buf;
-    buf = (int *) SUPERLU_MALLOC(n * sizeof(int));
+    buf = (int *) SUPERLU_MALLOC((size_t) n * sizeof(int));
     if ( !buf ) {
 	ABORT("SUPERLU_MALLOC fails for buf in intMalloc()");
     }
